change timestamp in imu-data from float to long

+++ annotell-input-api/annotell/input_api/model/ego/utils.py
@@ -1,1 +1,1 @@
-UnixTimestampNs = float  # A unix timestamp measured in nano seconds
+UnixTimestampNs = int  # A unix timestamp measured in nano seconds
